+++ files/zc_plugins/ProductsReviewReminder/v2.0.11/admin/includes/languages/english/extra_definitions/lang.products_review_reminder_menu.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'BOX_CONFIGURATION_REVIEW_REMINDER' => 'Products Review Reminder',
+    'BOX_TOOLS_REVIEW_REMINDER' => 'Products Review Reminder',
+];