+++ files/zc_plugins/ProductsReviewReminder/v2.0.11/catalog/includes/modules/pages/addon_my_reviews/main_template_vars.php
@@ -1,0 +1,5 @@
+<?php
+
+declare(strict_types=1);
+
+require dirname(__DIR__, 3) . '/templates/template_default/templates/tpl_addon_my_reviews_default.php';